chore: split stories
