+++ packages/brightcove-video/brightcove-video.stories.web.js
@@ -1,0 +1,34 @@
+import React from "react";
+import { View, Image, Text } from "react-native";
+
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { storiesOf } from "@storybook/react-native";
+import BrightcoveVideo from "./brightcove-video";
+
+const policyKey =
+  "BCpkADawqM0NK0Rq8n6sEQyWykemrqeSmIQqqVt3XBrdpl8TYlvqN3hwKphBJRnkPgx6WAbozCW_VgTOBCNf1AQRh8KnmXSXfveQalRc5-pyNlSod5XzP99If2U";
+const accountId = "57838016001";
+const videoId = "4084164751001";
+const posterImageURI =
+  "https://clips.news.co.uk/thetimes/p5dzhoYzE6kYmndXxYdBsfnnyMzDVTyD/Ut_HKthATH4eww8X4xMDoxOmFkOxyVqc";
+
+storiesOf(
+  "BrightcoveVideo (Web)",
+  module
+).add("Launcher with custom play icon defined by emoji", () => {
+  const playIconEmoji = () => (
+    <span role="img" aria-label="play-video">
+      ▶️
+    </span>
+  );
+
+  return (
+    <BrightcoveVideo
+      policyKey={policyKey}
+      videoId={videoId}
+      accountId={accountId}
+      poster={{ uri: posterImageURI }}
+      playIcon={playIconEmoji()}
+    />
+  );
+});
